Move pytrakt_extensions.py -> plex_trakt_sync/pytrakt_extensions.py
